"Finished conversion to cpp"

+++ source/server.cpp
@@ -3,7 +3,7 @@
 int main(int argc, char const *argv[]) {
 
     std::array<Card, CARD_NO> cards;
-    for (size_t i = 0; i < CARD_NO; i++) {
+    for (uint8_t i = 0; i < CARD_NO; i++) {
         cards[i] = Card(i);
     }
     shuffle(cards, *argv[1]);
